Separate versioning from the release plugin
