+++ buildSrc/src/main/java/ReleasePlugin.java
@@ -1,4 +1,3 @@
-import org.gradle.api.JavaVersion;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
@@ -6,84 +5,50 @@ import org.gradle.api.publish.PublishingExtension;
 import org.gradle.api.publish.maven.MavenPublication;
 import org.gradle.api.tasks.TaskProvider;
 import org.gradle.authentication.http.BasicAuthentication;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.UncheckedIOException;
-import java.nio.file.Files;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.EnumSet;
-import java.util.Locale;
-import java.util.Properties;
-import java.util.Set;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
 /**
  * Takes care of adding tasks and configurations to build developer distributions and releases.
  */
 public class ReleasePlugin implements Plugin<Project> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(ReleasePlugin.class);
-
-    private static final DateTimeFormatter SNAPSHOT_TIMESTAMP_FORMATTER = DateTimeFormatter.ofPattern("yyyyMMddHHmmssZ", Locale.US).withZone(ZoneOffset.UTC);
-    private static final String BUILD_RECEIPT_NAME = "build-receipt.properties";
-    private static final String BUILD_TIMESTAMP_PROPERTY = "buildTimestamp";
     private static final String UPLOAD_MAIN_TASK_NAME = "uploadMain";
     private static final String UPLOAD_JNI_TASK_NAME = "uploadJni";
     private static final String UPLOAD_NCURSES_JNI_TASK_NAME = "uploadNcursesJni";
-    private static final String ALPHA_VERSION_PROPERTY_NAME = "alpha";
 
     @Override
     public void apply(Project project) {
-        VersionDetails.BuildType buildType = determineBuildType(project);
-        VersionDetails versions = project.getExtensions().create("versions", VersionDetails.class, buildType);
-        if (project.hasProperty(ALPHA_VERSION_PROPERTY_NAME)) {
-            versions.setNextAlphaPostfix(project.property(ALPHA_VERSION_PROPERTY_NAME).toString());
-        }
-
-        if (buildType != VersionDetails.BuildType.Dev && JavaVersion.current() != JavaVersion.VERSION_1_8) {
-            throw new RuntimeException("Java 8 is required to build a release of native-platform. Later versions are not supported.");
-        }
-
-        String buildTimestamp = determineBuildTimestamp(project);
-        writeBuildTimestamp(buildTimestamp, project);
-
-        VersionCalculator versionCalculator = new VersionCalculator(versions, buildType, buildTimestamp);
-
+        Project rootProject = project.getRootProject();
+        rootProject.getPlugins().apply(VersioningPlugin.class);
+        VersionDetails versions = rootProject.getExtensions().getByType(VersionDetails.class);
         project.getGradle().getTaskGraph().whenReady(graph -> {
+            System.out.println(versions.getVersion());
             if (graph.hasTask(":" + UPLOAD_MAIN_TASK_NAME) || graph.hasTask(":" + UPLOAD_JNI_TASK_NAME) || graph.hasTask(":" + UPLOAD_NCURSES_JNI_TASK_NAME)) {
-                project.getLogger().lifecycle("##teamcity[buildStatus text='{build.status.text}, Published version {}']", versionCalculator);
+                project.getLogger().lifecycle("##teamcity[buildStatus text='{build.status.text}, Published version {}']", versions.getVersion());
             }
         });
 
-        project.allprojects(subproject -> {
-            subproject.getPlugins().apply(UploadPlugin.class);
-            subproject.getPlugins().apply(PublishPlugin.class);
-            subproject.setVersion(versionCalculator);
+        project.getPlugins().apply(UploadPlugin.class);
+        project.getPlugins().apply(PublishPlugin.class);
+        project.setVersion(versions.getVersion());
 
-            // Use authenticated snapshot/bintray repo while building a test distribution during snapshot/release
-            final BintrayCredentials credentials = subproject.getExtensions().getByType(BintrayCredentials.class);
+        // Use authenticated snapshot/bintray repo while building a test distribution during snapshot/release
+        final BintrayCredentials credentials = project.getExtensions().getByType(BintrayCredentials.class);
 
-            buildType.getReleaseRepository().ifPresent(releaseRepository -> {
-                credentials.assertPresent();
-                subproject.getRepositories().maven(repo -> {
-                    repo.setUrl(releaseRepository.getUrl());
-                    repo.getCredentials().setUsername(credentials.getUserName());
-                    repo.getCredentials().setPassword(credentials.getApiKey());
-                    repo.getAuthentication().create("basic", BasicAuthentication.class);
-                });
+        versions.getReleaseRepository().ifPresent(releaseRepository -> {
+            credentials.assertPresent();
+            project.getRepositories().maven(repo -> {
+                repo.setUrl(releaseRepository.getUrl());
+                repo.getCredentials().setUsername(credentials.getUserName());
+                repo.getCredentials().setPassword(credentials.getApiKey());
+                repo.getAuthentication().create("basic", BasicAuthentication.class);
             });
-
-            addUploadLifecycleTasks(subproject, buildType);
         });
+
+        addUploadLifecycleTasks(project, versions.getReleaseRepository());
     }
 
-    private void addUploadLifecycleTasks(Project project, VersionDetails.BuildType buildType) {
+    private void addUploadLifecycleTasks(Project project, Optional<VersionDetails.ReleaseRepository> releaseRepository) {
         Task uploadMainLifecycle = project.getTasks().maybeCreate(UPLOAD_MAIN_TASK_NAME);
         uploadMainLifecycle.setGroup("Upload");
         uploadMainLifecycle.setDescription("Upload Main publication");
@@ -99,9 +64,9 @@ public class ReleasePlugin implements Plugin<Project> {
         project.getExtensions().configure(
             PublishingExtension.class,
             extension -> extension.getPublications().withType(MavenPublication.class, publication -> {
-                String uploadTaskName = buildType.getReleaseRepository().map(releaseRepository ->
-                    releaseRepository.getType() == VersionDetails.RepositoryType.Maven
-                        ? BasePublishPlugin.publishTaskName(publication, releaseRepository.name())
+                String uploadTaskName = releaseRepository.map(repository ->
+                    repository.getType() == VersionDetails.RepositoryType.Maven
+                        ? BasePublishPlugin.publishTaskName(publication, repository.name())
                         : UploadPlugin.uploadTaskName(publication))
                     .orElse(BasePublishPlugin.publishTaskName(publication, BasePublishPlugin.LOCAL_FILE_REPOSITORY_NAME));
                 TaskProvider<Task> uploadTask = project.getTasks().named(uploadTaskName);
@@ -114,103 +79,5 @@ public class ReleasePlugin implements Plugin<Project> {
                     }
                 }
             }));
-    }
-
-    private void writeBuildTimestamp(String buildTimestamp, Project project) {
-        File buildReceiptFile = project.getRootProject().file(BUILD_RECEIPT_NAME);
-        try (OutputStream outputStream = Files.newOutputStream(buildReceiptFile.toPath())) {
-            Properties properties = new Properties();
-            properties.setProperty(BUILD_TIMESTAMP_PROPERTY, buildTimestamp);
-            properties.store(outputStream, null);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    private String determineBuildTimestamp(Project project) {
-        File buildReceipt = new File(project.getRootProject().file("incoming-distributions"), BUILD_RECEIPT_NAME);
-        if (project.hasProperty("ignoreIncomingBuildReceipt") || !buildReceipt.isFile()) {
-            return ZonedDateTime.now().format(SNAPSHOT_TIMESTAMP_FORMATTER);
-        }
-        Properties properties = new Properties();
-        try (InputStream inputStream = Files.newInputStream(buildReceipt.toPath())) {
-            properties.load(inputStream);
-            String buildTimestamp = properties.getProperty(BUILD_TIMESTAMP_PROPERTY);
-            LOGGER.warn("Using build timestamp from incoming build receipt: {}", buildTimestamp);
-            return properties.getProperty("buildTimestamp");
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    private VersionDetails.BuildType determineBuildType(Project project) {
-        boolean release = project.hasProperty("release");
-        boolean milestone = project.hasProperty("milestone");
-        boolean alpha = project.hasProperty(ALPHA_VERSION_PROPERTY_NAME);
-        boolean snapshot = project.hasProperty("snapshot");
-
-        Set<VersionDetails.BuildType> enabledBuildTypes = EnumSet.noneOf(VersionDetails.BuildType.class);
-        if (release) {
-            enabledBuildTypes.add(VersionDetails.BuildType.Release);
-        }
-        if (milestone) {
-            enabledBuildTypes.add(VersionDetails.BuildType.Milestone);
-        }
-        if (alpha) {
-            enabledBuildTypes.add(VersionDetails.BuildType.Alpha);
-        }
-        if (snapshot) {
-            enabledBuildTypes.add(VersionDetails.BuildType.Snapshot);
-        }
-        if (enabledBuildTypes.size() > 1) {
-            throw new UnsupportedOperationException(
-                "Cannot build " +
-                    enabledBuildTypes.stream()
-                        .map(Object::toString)
-                        .collect(Collectors.joining(" and ")) +
-                    " in same build.");
-        }
-        return enabledBuildTypes.stream().findFirst().orElse(VersionDetails.BuildType.Dev);
-    }
-
-    private static class VersionCalculator {
-        private final VersionDetails release;
-        private final VersionDetails.BuildType buildType;
-        private final String buildTimestamp;
-        private String version;
-
-        VersionCalculator(VersionDetails release, VersionDetails.BuildType buildType, String buildTimestamp) {
-            this.release = release;
-            this.buildType = buildType;
-            this.buildTimestamp = buildTimestamp;
-        }
-
-        @Override
-        public String toString() {
-            if (version == null) {
-                String nextVersion = release.getNextVersion();
-                if (nextVersion == null) {
-                    throw new UnsupportedOperationException("Next version not specified.");
-                }
-                if (buildType == VersionDetails.BuildType.Release) {
-                    version = nextVersion;
-                } else if (buildType == VersionDetails.BuildType.Milestone) {
-                    if (release.getNextSnapshot() == null) {
-                        throw new UnsupportedOperationException("Next milestone not specified.");
-                    }
-                    version = nextVersion + "-milestone-" + release.getNextSnapshot();
-                } else if (buildType == VersionDetails.BuildType.Alpha) {
-                    if (release.getNextAlphaPostfix() == null || release.getNextAlphaPostfix().isEmpty()) {
-                        throw new UnsupportedOperationException("Next alpha version postfix not specified.");
-                    }
-                    version = nextVersion + "-" + release.getNextAlphaPostfix();
-                } else if (buildType == VersionDetails.BuildType.Snapshot) {
-                    version = nextVersion + "-snapshot-" + buildTimestamp;
-                } else {
-                    version = nextVersion + "-dev";
-                }
-            }
-            return version;
-        }
     }
 }
